% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelinesScripts.R
\name{DADA2pipe}
\alias{DADA2pipe}
\title{DADA2 pipeline function}
\usage{
DADA2pipe(
  in.dir,
  ref,
  ID,
  truncLen = c(200, 150),
  trimLeft = c(18, 19),
  maxN = 0,
  maxEE = c(2, 2),
  truncQ = 2,
  rm.phix = TRUE,
  compress = TRUE,
  multithread = TRUE
)
}
\arguments{
\item{in.dir}{directory path to fastq files}

\item{ref}{directory path to reference for taxomony annotation}

\item{ID}{string to search file names for sample ID values}

\item{truncLen}{where to truncate sequence forward and reverse reads. Default value is c(200, 150). See DADA2 documentation}

\item{trimLeft}{removes bases from left, used to remove primers left over. Default value is c(18, 19). See DADA2 documentation}

\item{maxN}{sets maximum number of N values in sequences that are retained. Default value is 0. See DADA2 documentation}

\item{maxEE}{sets the maximum number of expected errors for each read. Default value is c(2,2). See DADA2 documentation}

\item{truncQ}{paramter to tune sequence truncation. Default is 2. See DADA2 documentation}

\item{rm.phix}{logical to remove PhiX. Default is TRUE. See DADA2 documentation}

\item{compress}{logical. Default is TRUE. See DADA2 documentation}

\item{multithread}{logical. Default is TRUE. See DADA2 documentation}

\item{metadf}{metadata for sequence annotation}
}
\description{
This function implements the DADA2 sequence QC and annotation pipeline for amplicon sequence libraries. Requires a directory of fastq files that are named with _R1_001.fastq.gz or _R2_001.fastq.gz, formatted with searchable sample IDs incorporated into the file names. Depends on stringr and DADA2. Returns a phyloseq object with metadata.
}
\examples{
DAD2Apipe()
}
\keyword{DADA2}
\keyword{pipeline}
